Use dwc2 gadget mode with serial and networking
This creates a usb0 ethernet device on the pi zero, and the host
computer should see a CDC Composite Gadget ethernet device. It also
creates a /dev/ttyGS0 serial device on the pi zero and a serial device
on the host computer. If you switch erlinit to use /dev/ttyGS0, you can
connect to the iex console over serial with a baud rate of 115200.